pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dv-regression-lab"
version = "0.1.0"
description = "Local-first RTL/DV regression control tower with triage, flaky detection, and artifact traceability."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Doeon Kim" }]
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27",
  "pytest>=8.0",
  "ruff>=0.11",
]

[project.scripts]
dvrl = "dv_regression_lab.cli:main"

[tool.setuptools.packages.find]
include = ["dv_regression_lab*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

